pyproject.toml
[project]
name = "rag-chatbot-hw-car"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.14.0",
    "pinecone>=8.0.0",
    "python-dotenv>=1.2.1",
]
